+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## unreleased
 
+## 1.5.0
+
+* Added `leak-detection-threshold` option for Hikari config.
+* Added `register-mbeans` option for Hikari config.
+
+## 1.4.0
+
 * Added `sqlite` support with `org.sqlite.JDBC` adapter
 * Added a general purpose way to alter Hikari config